requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-CORS
python-dotenv
psycopg2-binary
